refactor: split pack_kernel_args out of gpu.rs, and use is_multiple_of
Two CI failures caught BEFORE the campaign rather than after it, which is the
whole point:

  * `gpu.rs` went 479 -> 512 lines when `pack_kernel_args` landed, crossing the
    repo's per-file cap. Moved to `gpu_args.rs` (470 + 54). It earns its own file
    anyway: the packing is the one place that knows an argument may span MORE
    than one slot, and the version it replaced silently truncated anything wider
    than 8 bytes.
  * `clippy::manual_is_multiple_of` on the TMA guard's `qk_stride % 8 != 0`.

Exact piecewise move, no behaviour change. fmt, LoC cap, clippy --workspace
--tests and cargo test --workspace all clean at this commit.

★ Learned the expensive way on #702: two style commits pushed AFTER its campaign
invalidated all ten records ("invalidated by crates/spark-model/src/layers/ops.rs")
and cost the whole 3.5-hour run. Static checks come before campaigns now.
